requirements.txt
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
pillow==10.4.0
numpy==1.26.4
pylint==3.2.6
pytest==8.3.2